+++ Src/Library/Endpoint/Endpoint.Setup.cs
@@ -9,6 +9,7 @@ namespace FastEndpoints;
 
 public abstract partial class Endpoint<TRequest, TResponse> : BaseEndpoint where TRequest : notnull, new() where TResponse : notnull
 {
+    private static readonly Type tRequest = typeof(TRequest);
     private static readonly Type tResponse = typeof(TResponse);
     private static readonly ConstructorInfo? rspDTOCtor = tResponse.GetConstructor(Type.EmptyTypes);