+++ Src/Library/changelog.md
@@ -1,2 +1,3 @@
 ### IMPROVEMENTS
 - optimize default request binder
+- global request binding modifier func support #202